Car Rental Webpage
Included the zip file along with all the files/folders.

This is a final project that focused on the use of JavaScript and forms. The site includes pages related to
cars and rental information with images as well as a registration page that allows the user to enter
their name, email, and their choice of car to rent. When submitted, the information is listed below,
and the user can print or save the document.